requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiofiles==24.1.0
